requirements.txt
apturl==0.5.2
asgiref==3.5.2
bcrypt==3.2.0
black==22.10.0
bleach==5.0.1
blinker==1.4
boto3==1.26.118
botocore==1.29.118
Brlapi==0.8.3
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cssselect2==0.7.0
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
Django==4.1.3
django-analytical==3.1.0
django-autoslug==1.9.8
django-cors-headers==3.13.0
django-filter==23.2
django-js-asset==2.0.0
django-meta==2.1.0
django-model-utils==4.2.0
django-mptt==0.14.0
django-storages==1.13.2
django-summernote==0.8.20.0
django-tenants==3.4.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
duplicity==0.8.21
fasteners==0.14.1
filelock==3.12.0
fonttools==4.37.4
future==0.18.2
html5lib==1.1
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
jeepney==0.7.1
jmespath==1.0.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
mypy-extensions==0.4.3
nepali-date==2.0.5
nepali-datetime==1.0.7
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
pathspec==0.10.2
pexpect==4.8.0
phonenumbers==8.12.56
Pillow==9.3.0
platformdirs==2.5.4
protobuf==3.12.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydyf==0.5.0
PyGObject==3.42.1
PyJWT==2.6.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyotp==2.7.0
pyparsing==2.4.7
pyphen==0.13.0
pyRFC3339==1.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-decouple==3.6
pytz==2022.6
pyxdg==0.27
PyYAML==5.4.1
qrcode==7.3.1
reportlab==3.6.8
requests==2.28.1
s3transfer==0.6.0
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.3
stripe==4.2.0
systemd-python==234
tinycss2==1.2.1
tomli==2.0.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.13
usb-creator==0.3.7
virtualenv==20.23.0
wadllib==1.3.6
weasyprint==57.0
webencodings==0.5.1
xdg==5
xkit==0.0.0
zipp==1.0.0
zopfli==0.2.1
